# src/swagger.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: Lawyer's Hub
  description: Hub where lawyer and client meets.
  
host: localhost:1802
basePath: /api
schemes:
  - http
paths:
  /common/getstates:
    get:
      summary: Return states. 
      description: Return all country states.
      operationId: getstates
      tags:
        - Common API
      produces:
        - application/json
      responses:
        200:
          description: OK
  /common/getcountry:
    get:
      summary: Return countries.
      description: return all country list as JSON
      operationId: getcountry
      tags:
        - Common API
      produces:
        - application/json
      responses:
        200:
          description: OK
  /common/getcities/{state}:
    get:
      summary: reutrn cities.
      operationId: getcities.
      tags:
        - Common API
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: state
          required: true
          type: string
          description: require state as input parameter.
      responses:
        200:
          description: Cities array.
          schema:
            type: object

  /v1/login:
    post:
      summary: user can login.
      operationId: login
      description: User can pass credentials and got token and user identity.
      tags:
        - Account
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: formData
          name: username
          type: string
          required: true
          description: Username is required.
        - in: formData
          name: password
          type: string
          required: true
          description: Password is required.
      responses:
        200:
          description: OK.
        404:
          description: Not Found.
        403:
          description: Forbidden Request.
  /v1/signup:
    post:
      summary: New user can signup.
      operationId: signup
      description: new user can signup giving his/her following information.
      tags:
        - Account
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          definition: create new user.
          schema:
            $ref: "#/definitions/signup"
      responses:
        200:
          description: OK
        404:
          description: Not Created.
  /v1/forgotpassword/{email}:
    get:
      summary: password reset link.
      operationId: forgotpassword
      description: User will get password reset link on its register email address.
      tags:
        - Account
      consumes:
        - application/json
      parameters:
        - in: path
          name: email
          type: string
          required: true
          description: Email Address.
      responses:
        200:
          description: OK
        400:
          description: Bad request.
  
  /v1/getAllClients/{userId}:
    get:
      summary: return all clients.
      description: return all the clients private and public as JSON.
      operationId: getAllClients
      tags:
        - Auth User
      produces:
        - appication/json
      parameters:
        - in: header
          name: authorization
          type: string
          required: true
          description: authorization token.
        - in: path
          name: userId
          type: string
          required: true
          description: userid
      responses:
        200:
          definition: OK
        404:
          definition: No client.
        401:
          defination: unauthorize request.
  /v1/getClientList/{userId}/{isprivate}:
    get:
      summary: Fetch registred user's client.
      description: Fetch registred user's client by userid and client type.
      operationId: getClientList
      tags:
        - Auth User
      parameters:
        - in: header
          name: authorization
          type: string
          required: true
          description: bearer token.
        - in: path
          name: userId
          type: string
          required: true
          description: userid
        - in: path
          name: isprivate
          type: boolean
          required: true
          description: Is client private.
      responses:
        200:
          description: OK
        404:
          description: No client.
  /v1/addClient:
    post:
      summary: User can add clients.
      description: user can add client, providing following details.
      operationId: addClient
      tags:
        - Auth User
      parameters:
        - in: header
          name: authorization
          type: string
          required: true
          description: bearer token
        - in: body
          name: addClient
          definition: add new client
          schema:
            $ref: "#/definitions/addclient"
      responses:
        200:
          description: OK

definitions:
  signup:
    type: object
    required:
      - firstName
      - lastName
      - phone
      - email
      - username
      - password
    properties:
      firstName:
        type: string
      midleName:
        type: string
      lastName:
        type: string
      phone:
        type: string
      email:
        type: string
      username:
        type: string
      password:
        type: string
  addclient:
    type: object
    required:
      - userid
      - firstName
      - lastName
      - address1
      - address2
      - country
      - state
      - district
      - city
      - pincode
      - email
      - phone
      - purpose
      - occupation
      - about
      - isprivate
    properties:
      userid:
        type: string
      firstName:
        type: string
      middleName:
        type: string
      lastName:
        type: string
      address1:
        type: string
      address2:
        type: string
      country:
        type: string
      state:
        type: string
      district:
        type: string
      city:
        type: string
      pincode:
        type: string
      email:
        type: string
      phone:
        type: string
      purpose:
        type: integer
      occupation:
        type: string
      about:
        type: string
      isprivate:
        type: boolean
     